changing the amount field back to integer

+++ crowdfunding/projects/models.py
@@ -70,7 +70,7 @@ class Project(models.Model):
         super().save(**kwargs)
 
 class Pledge(models.Model):
-    amount = models.DecimalField(max_digits=5, decimal_places=2)
+    amount = models.IntegerField()
     comment = models.CharField(max_length=200)
     anonymous = models.BooleanField()
     project = models.ForeignKey('Project', on_delete=models.CASCADE, related_name='pledges')
@@ -79,3 +79,6 @@ class Pledge(models.Model):
         on_delete=models.CASCADE,
         related_name='pledges',
     )
+
+
+
